+++ packages/api/endpoints/distribution.js
@@ -1,50 +1,25 @@
 'use strict';
 
-const get = require('lodash.get');
-const log = require('@cumulus/common/log');
-const { aws } = require('@cumulus/common');
+const urljoin = require('url-join');
+const { s3 } = require('@cumulus/common/aws');
 const { URL } = require('url');
+const { Cookie } = require('tough-cookie');
 
 const EarthdataLoginClient = require('../lib/EarthdataLogin');
-const {
-  OAuth2AuthenticationFailure
-} = require('../lib/OAuth2');
 
-class UnparsableGranuleLocationError extends Error {
-  constructor(granuleLocation) {
-    super(`Granule location "${granuleLocation}" could not be parsed`);
+const { RecordDoesNotExist } = require('../lib/errors');
+const { AccessToken } = require('../models');
+const {
+  NotFoundResponse,
+  TemporaryRedirectResponse
+} = require('../lib/responses');
+const { findCaseInsensitiveValue } = require('../lib/utils');
+
+class UnparsableFileLocationError extends Error {
+  constructor(fileLocation) {
+    super(`File location "${fileLocation}" could not be parsed`);
     this.name = this.constructor.name;
   }
-}
-
-/**
- * Build an API Gateway redirect response
- *
- * @param {string} url - the URL to redirect to
- * @returns {Object} an API Gateway response object
- */
-function buildRedirectResponse(url) {
-  return {
-    statusCode: 302,
-    body: 'Redirect',
-    headers: {
-      Location: url,
-      'Strict-Transport-Security': 'max-age=31536000'
-    }
-  };
-}
-
-/**
- * Build an API Gateway client error response
- *
- * @param {string} errorMessage - the error message to be returned in the response
- * @returns {Object} an API Gateway response object
- */
-function buildClientErrorResponse(errorMessage) {
-  return {
-    statusCode: 400,
-    body: JSON.stringify({ error: errorMessage })
-  };
 }
 
 /**
@@ -54,17 +29,17 @@ function buildClientErrorResponse(errorMessage) {
  * @returns {Object} - bucket/key in the form of
  * { Bucket: x, Key: y }
  */
-function getBucketAndKeyFromPathParams(pathParams) {
+function getFileBucketAndKey(pathParams) {
   const fields = pathParams.split('/');
 
   const Bucket = fields.shift();
   const Key = fields.join('/');
 
   if (Bucket.length === 0 || Key.length === 0) {
-    throw new UnparsableGranuleLocationError(pathParams);
+    throw new UnparsableFileLocationError(pathParams);
   }
 
-  return { Bucket, Key };
+  return [Bucket, Key];
 }
 
 /**
@@ -76,7 +51,7 @@ function getBucketAndKeyFromPathParams(pathParams) {
  * @param {string} username - the username to add to the redirect url
  * @returns {string} a URL
  */
-function getSignedUrl(s3Client, Bucket, Key, username) {
+function getSignedS3Url(s3Client, Bucket, Key, username) {
   const signedUrl = s3Client.getSignedUrl('getObject', { Bucket, Key });
 
   const parsedSignedUrl = new URL(signedUrl);
@@ -85,94 +60,131 @@ function getSignedUrl(s3Client, Bucket, Key, username) {
   return parsedSignedUrl.toString();
 }
 
-/**
- * Given a an API Gateway request, return either the proxy path parameter or
- *   the state query string parameter
- *
- * @param {Object} request - an API Gatway request object
- * @returns {string|undefined} a granule location
- */
-function getGranuleLocationFromRequest(request) {
-  return get(request, 'pathParameters.proxy')
-    || get(request, 'queryStringParameters.state');
+function getAccessTokenFromRequest(request) {
+  const cookieHeaders = findCaseInsensitiveValue(request.multiValueHeaders || {}, 'Cookie') || [];
+  const cookies = cookieHeaders.map(Cookie.parse);
+  const accessTokenCookie = cookies.find((c) => c.key === 'accessToken');
+
+  return accessTokenCookie ? accessTokenCookie.value : null;
 }
 
-/**
- * Return the username associated with an OAuth2 authorization code
- *
- * @param {EarthdataLoginClient} earthdataLoginClient - an Earthdata Login Client
- * @param {string} authorizationCode - the OAuth2 authorization code to use
- * @returns {string} an Earthdata username
- */
-async function getUsernameFromAuthorizationCode(earthdataLoginClient, authorizationCode) {
+function isRedirectRequest(request) {
+  return request.resource === '/redirect';
+}
+
+function isAccessTokenExpired(accessTokenRecord) {
+  return accessTokenRecord.expirationTime < Date.now();
+}
+
+async function handleRedirectRequest(params = {}) {
   const {
-    username
-  } = await earthdataLoginClient.getAccessToken(authorizationCode);
+    accessTokenModel,
+    authClient,
+    distributionUrl,
+    request
+  } = params;
 
-  return username;
+  const { code, state } = request.queryStringParameters;
+
+  const getAccessTokenResponse = await authClient.getAccessToken(code);
+
+  await accessTokenModel.create({
+    accessToken: getAccessTokenResponse.accessToken,
+    expirationTime: getAccessTokenResponse.expirationTime,
+    refreshToken: getAccessTokenResponse.refreshToken,
+    username: getAccessTokenResponse.username
+  });
+
+  return new TemporaryRedirectResponse({
+    location: urljoin(distributionUrl, state),
+    cookies: [
+      new Cookie({
+        key: 'accessToken',
+        value: getAccessTokenResponse.accessToken,
+        expires: new Date(getAccessTokenResponse.expirationTime),
+        httpOnly: true,
+        secure: true
+      })
+    ]
+  });
 }
 
-/**
- * An AWS API Gateway function that either requests authentication,
- * or if authentication is found then redirects to an S3 file for download
- *
- * @param {Object} request - an API Gateway request object
- * @param {EarthdataLoginClient} earthdataLoginClient - an instance of an
- *   EarthdataLoginClient that will be used when authorizing this request
- * @param {AWS.S3} s3Client - an AWS S3 client that will be used when processing
- *   this request
- * @returns {Promise<Object>} an API Gateway response object
- */
-async function handleRequest(request, earthdataLoginClient, s3Client) {
-  const granuleLocation = getGranuleLocationFromRequest(request);
+async function handleFileRequest(params = {}) {
+  const {
+    accessTokenModel,
+    authClient,
+    request,
+    s3Client
+  } = params;
 
-  const authorizationCode = get(request, 'queryStringParameters.code');
+  const redirectToGetAuthorizationCode = new TemporaryRedirectResponse({
+    location: authClient.getAuthorizationUrl(request.pathParameters.proxy)
+  });
 
-  if (authorizationCode) {
-    try {
-      const username = await getUsernameFromAuthorizationCode(
-        earthdataLoginClient,
-        authorizationCode
-      );
+  const accessToken = getAccessTokenFromRequest(request);
 
-      const {
-        Bucket,
-        Key
-      } = getBucketAndKeyFromPathParams(granuleLocation);
+  if (!accessToken) return redirectToGetAuthorizationCode;
 
-      log.info({
-        username,
-        accessDate: Date.now(),
-        bucket: Bucket,
-        file: Key,
-        sourceIp: get(request, 'requestContext.identity.sourceIp')
-      });
-
-      const s3RedirectUrl = getSignedUrl(
-        s3Client,
-        Bucket,
-        Key,
-        username
-      );
-
-      return buildRedirectResponse(s3RedirectUrl);
+  let accessTokenRecord;
+  try {
+    accessTokenRecord = await accessTokenModel.get({ accessToken });
+  }
+  catch (err) {
+    if (err instanceof RecordDoesNotExist) {
+      return redirectToGetAuthorizationCode;
     }
-    catch (err) {
-      if (err instanceof OAuth2AuthenticationFailure) {
-        return buildClientErrorResponse('Failed to get EarthData token');
-      }
 
-      if (err instanceof UnparsableGranuleLocationError) {
-        return buildClientErrorResponse(err.message);
-      }
-
-      throw err;
-    }
+    throw err;
   }
 
-  const authorizationUrl = earthdataLoginClient.getAuthorizationUrl(granuleLocation);
+  if (isAccessTokenExpired(accessTokenRecord)) {
+    return redirectToGetAuthorizationCode;
+  }
 
-  return buildRedirectResponse(authorizationUrl);
+  let fileBucket;
+  let fileKey;
+  try {
+    [fileBucket, fileKey] = getFileBucketAndKey(request.pathParameters.proxy);
+  }
+  catch (err) {
+    if (err instanceof UnparsableFileLocationError) return new NotFoundResponse();
+    throw err;
+  }
+
+  const signedS3Url = getSignedS3Url(
+    s3Client,
+    fileBucket,
+    fileKey,
+    accessTokenRecord.username
+  );
+
+  return new TemporaryRedirectResponse({ location: signedS3Url });
+}
+
+async function handleRequest(params = {}) {
+  const {
+    accessTokenModel,
+    authClient,
+    distributionUrl,
+    request,
+    s3Client
+  } = params;
+
+  if (isRedirectRequest(request)) {
+    return handleRedirectRequest({
+      accessTokenModel,
+      authClient,
+      distributionUrl,
+      request
+    });
+  }
+
+  return handleFileRequest({
+    accessTokenModel,
+    authClient,
+    request,
+    s3Client
+  });
 }
 
 /**
@@ -189,9 +201,13 @@ async function handleApiGatewayRequest(event) {
     redirectUri: process.env.DEPLOYMENT_ENDPOINT
   });
 
-  const s3Client = aws.s3();
-
-  return handleRequest(event, earthdataLoginClient, s3Client);
+  return handleRequest({
+    accessTokenModel: new AccessToken(),
+    authClient: earthdataLoginClient,
+    distributionUrl: process.env.DISTRIBUTION_URL,
+    request: event,
+    s3Client: s3()
+  });
 }
 
 module.exports = {