+++ packages/api/lib/utils.js
@@ -77,6 +77,32 @@ function getGranuleProductVolume(granuleFiles) {
   return fileSizes.reduce((a, b) => a + b);
 }
 
+/**
+ * Find a property name in an object in a case-insensitive manner
+ *
+ * @param {Object} obj - the object to be searched
+ * @param {string} keyArg - the name of the key to find
+ * @returns {string|undefined} - the name of the matching key, or undefined if
+ *   none was found
+ */
+function findCaseInsensitiveKey(obj, keyArg) {
+  const keys = Object.keys(obj);
+  return keys.find((key) => key.toLowerCase() === keyArg.toLowerCase());
+}
+exports.findCaseInsensitiveKey = findCaseInsensitiveKey;
+
+/**
+ * Find a property value in an object in a case-insensitive manner
+ *
+ * @param {Object} obj - the object to be searched
+ * @param {string} keyArg - the name of the key to find
+ * @returns {*} the matching value
+ */
+function findCaseInsensitiveValue(obj, keyArg) {
+  return obj[findCaseInsensitiveKey(obj, keyArg)];
+}
+exports.findCaseInsensitiveValue = findCaseInsensitiveValue;
+
 module.exports.errorify = errorify;
 module.exports.parseException = parseException;
 module.exports.deconstructCollectionId = deconstructCollectionId;